Fix value_codec test failure
- Commented out test causing undetermined values.
- clang on MacOS and gcc on Linux give different results when optimization
is turned on. Per discussion on StackOverflow:

"A prvalue of a floating-point type can be converted to a prvalue of an
integer type. The conversion truncates; that is, the fractional part is
discarded. The behavior is undefined if the truncated value cannot be
represented in the destination type.

So implementations can do whatever they like for your case, as the
behavior is undefined."